add config and endpoints to Deposit, update classes to use embedded object of User and House id (UserHouse)

+++ src/main/java/eeet2582/realestatemgt/house/HouseService.java
@@ -26,6 +26,9 @@ public class HouseService {
     }
 
     public void deleteById(Long houseId) {
+        if (!houseRepository.existsById(houseId))
+            throw new IllegalStateException("House with houseId=" + houseId + " does not exist!");
+
         houseRepository.deleteById(houseId);
     }
 }
